requirements.txt
telethon>=1.36.0
apscheduler>=3.10.0
pytz>=2024.1
python-dotenv>=1.0.0
qrcode>=7.4.2
